test 대채 앱 실행 테스트
